moved image modal up into main page
