+++ frontend/src/components/Profile/Profile.js
@@ -5,7 +5,11 @@ import Post from "../Post/Post";
 import { user } from "../../test-data/user-data";
 import { posts } from "../../test-data/post-data";
 import { AiOutlinePlus } from "react-icons/ai";
+import { useOutletContext } from "react-router-dom";
+
 const Profile = () => {
+  const [handleImageClick] = useOutletContext();
+
   // Fetch data using route path and params
   return (
     <div className={styles["profile-container"]}>
@@ -16,7 +20,7 @@ const Profile = () => {
         bio={user.bio}
       />
       {posts.map((post, index) => (
-        <Post key={index} post={post} />
+        <Post key={index} post={post} handlePostImageClick={handleImageClick} />
       ))}
       {/**TODO: create pop up modal */}
       <button className={styles["create-post"]}>
